# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sistema de Matrículas",
    "version": "1.0.0",
    "description": "API para SPA de matrículas - documentação gerada via swagger-jsdoc"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "ra": {
            "type": [
              "string",
              "null"
            ]
          },
          "siape": {
            "type": [
              "string",
              "null"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AuthRequest": {
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/Usuario"
          }
        }
      },
...
      "Matricula": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "aluno_id": {
            "type": "integer"
          },
          "turma_id": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MatriculaCreate": {
        "type": "object",
        "required": [
          "turmaId"
        ],
        "properties": {
          "turmaId": {
            "type": "integer"
          }
        }
      },
      "Disciplina": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "carga_horaria": {
            "type": "integer"
          }
        },
        "required": [
          "codigo",
          "nome"
        ]
      },
      "DisciplinaCreate": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "carga_horaria": {
            "type": "integer"
          }
        },
        "required": [
          "codigo",
          "nome"
        ]
      },
      "Horario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "dia": {
            "type": "integer",
            "description": "1..7 (1=domingo)"
          },
          "turno": {
            "type": "integer",
            "description": "1=manhã,2=tarde,3=noite"
          },
          "codigo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          }
        },
        "required": [
          "dia",
          "turno",
          "codigo"
        ]
      },
      "HorarioCreate": {
        "type": "object",
        "properties": {
          "dia": {
            "type": "integer"
          },
          "turno": {
            "type": "integer"
          },
          "codigo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          }
        },
        "required": [
          "dia",
          "turno",
          "codigo"
        ]
      }
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autenticar usuário e receber access + refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retorna accessToken e refreshToken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/disciplinas": {
      "get": {
        "tags": [
          "Disciplinas"
        ],
        "summary": "Lista todas as disciplinas",
        "responses": {
          "200": {
            "description": "Lista de disciplinas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Disciplina"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Disciplinas"
        ],
        "summary": "Cria uma nova disciplina",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisciplinaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Disciplina criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Disciplina"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação"
          },
          "409": {
            "description": "Conflito (código de disciplina já existe)"
          }
        }
      }
    },
    "/api/disciplinas/{id}": {
      "put": {
        "tags": [
          "Disciplinas"
        ],
        "summary": "Atualiza uma disciplina",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID da disciplina"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisciplinaCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Disciplina atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Disciplina"
                }
              }
            }
          },
          "409": {
            "description": "Conflito (código de disciplina já existe)"
          }
        }
      }
    },
    "/api/horarios": {
      "get": {
        "tags": [
          "Horarios"
        ],
        "summary": "Lista todos os horários",
        "responses": {
          "200": {
            "description": "Lista de horários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Horario"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Horarios"
        ],
        "summary": "Cria um novo horário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HorarioCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Horário criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Horario"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação"
          },
          "409": {
            "description": "Conflito (código de horário já existe)"
          }
        }
      }
    },
    "/api/horarios/{id}": {
      "put": {
        "tags": [
          "Horarios"
        ],
        "summary": "Atualiza um horário",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID do horário"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HorarioCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Horário atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Horario"
                }
              }
            }
          },
          "409": {
            "description": "Conflito (código de horário já existe)"
          }
        }
      }
    },
    "/api/matriculas": {
      "post": {
        "tags": [
          "Matriculas"
        ],
        "summary": "Matricular aluno em uma turma (requer autenticação como ALUNO)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatriculaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Matrícula criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Matricula"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar usuários (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários"
          }
        }
      }
    },
...
  },
...
}